LayoutTests/platform/ios-sim-deprecated/fast/forms/input-widths-expected.txt

Text field types should have the same default widths.

FAIL document.getElementById("email").offsetWidth should be 135. Was 136.
PASS document.getElementById("number").offsetWidth is baseWidth
PASS document.getElementById("password").offsetWidth is baseWidth
FAIL document.getElementById("tel").offsetWidth should be 135. Was 136.
PASS document.getElementById("url").offsetWidth is baseWidth
PASS successfullyParsed is true

TEST COMPLETE
